pyproject.toml
[tool.poetry]
name = "rslv"
version = "0.1.0"
description = "Provides an identifier resolver service in FastAPI."
authors = ["datadave <605409+datadavev@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0"
pydantic-settings = "^2"
sqlalchemy = "^2.0.20"
#jinja2 = "^3.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7"
requests = "^2.31.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}

[tool.poetry.scripts]
rslv = "scripts.manage:main"
crne = "scripts.crne:main"
compare_ezid_n2t_naans = "scripts.compare_ezid_n2t_naans:main"

[tool.poetry.group.cli.dependencies]
click = "^8.1.7"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
